#42 : moved slackNameRequest and UserDTO into package DTO
