+++ src/main/java/juja/microservices/gamification/slackbot/dao/impl/RestUserRepository.java
@@ -7,8 +7,8 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import juja.microservices.gamification.slackbot.dao.UserRepository;
 import juja.microservices.gamification.slackbot.exceptions.GamificationExchangeException;
 import juja.microservices.gamification.slackbot.exceptions.UserNotFoundException;
-import juja.microservices.gamification.slackbot.model.SlackNameRequest;
-import juja.microservices.gamification.slackbot.model.UserDTO;
+import juja.microservices.gamification.slackbot.model.DTO.SlackNameRequest;
+import juja.microservices.gamification.slackbot.model.DTO.UserDTO;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.*;
 import org.springframework.stereotype.Repository;
